Refactor into modular packages and modern UI helpers
