+++ db/__init__.py
@@ -1,0 +1,3 @@
+from .manager import DatabaseManager
+
+__all__ = ["DatabaseManager"]
